Split steps.go into smaller focused files
Refactor pkg/schema/steps.go (900 lines) into:
- steps.go (299 lines): Core Step wrapper, unmarshaling, common utilities
- run_step.go (186 lines): RunStep struct and run-related types
- build_step.go (273 lines): BuildStep, BuildConfig, and build types
- watch.go (145 lines): Watch struct and watch-related methods

This separation improves code organization by grouping related types
together while maintaining full functionality and test coverage.
